refactor: extract composer draft hook
